Initial commit.
This repository is designed to keep available tiny & useful pieces of code which are to be employed to solve tiny routine tasks fast and with no manual actions.

For now, it contains RuNet domains grabber script for purpose of collecting all available domains in russian segment of Internet. One can find such a dataset useful for future research, such as linked graph reconstruction.
Grabber imports Selenium as the most popular web scraping instrument.

A service provided by ruCaptcha.com is also used to resolve captcha so we could collect all the data to a local machine. More info: https://rucaptcha.com/api-rucaptcha Tesseract have to be engaged for this purpose in further as soon as captcha used on the site is not very hard to resolve by machine learning techniques.